requirements.txt
babel==2.9.1
fastapi==0.75.0
markdown-builder==0.1.2
money==1.3.0
mypy==0.940
peewee==3.14.10
psycopg2==2.9.3
pycodestyle==2.8.0
pydantic==1.9.0
python-dotenv==0.19.2
python-telegram-bot==13.11
requests==2.27.1
types-requests==2.27.11
uvicorn[standard]==0.17.6
